src/requirements.txt
pyserial
minimalmodbus
